backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
aiofiles
ultralytics
opencv-python-headless
httpx
python-dotenv
alembic
google-generativeai
